Remove binding errors on startup
This improves startup and removes debug messages. Turns out the TrayIcon
control will set itself as the DataContext if none is set, which causes
errors because the Binding paths don't exist on it. There was no visible
impact to users since DataContext was immediately being set, but it
polluted the debug log and slightly slowed down startup.

This refactor extract a xaml binadable class that is now set as the
DataContext from xaml to avoid these issues.